+++ Kernel/Output/HTML/Dashboard/TicketGeneric.pm
@@ -38,6 +38,13 @@ sub new {
         $Self->{$Item} = $ParamObject->GetParam( Param => $Item ) || $Param{$Item};
     }
 
+    # Get add filters param.
+    $Self->{AddFilters} = $ParamObject->GetParam( Param => 'AddFilters' ) || $Param{AddFilters} || 0;
+    $Self->{TabAction}  = $ParamObject->GetParam( Param => 'TabAction' )  || $Param{TabAction}  || 0;
+
+    # Get previous sorting column.
+    $Self->{SortingColumn} = $ParamObject->GetParam( Param => 'SortingColumn' ) || $Param{SortingColumn};
+
     # set filter settings
     for my $Item (qw(ColumnFilter GetColumnFilter GetColumnFilterSelect)) {
         $Self->{$Item} = $Param{$Item};
@@ -577,8 +584,40 @@ sub Run {
     );
     $CacheKey .= '-' . $CacheColumns if $CacheColumns;
 
-    $CacheKey .= '-' . $Self->{SortBy}  if defined $Self->{SortBy};
-    $CacheKey .= '-' . $Self->{OrderBy} if defined $Self->{OrderBy};
+    # If SortBy parameter is not defined, set to value from %TicketSearch, otherwise set to default value 'Age'.
+    if ( !defined $Self->{SortBy} ) {
+        if ( defined $TicketSearch{SortBy} && $Self->{ValidSortableColumns}->{ $TicketSearch{SortBy} } ) {
+            $Self->{SortBy} = $TicketSearch{SortBy};
+        }
+        else {
+            $Self->{SortBy} = 'Age';
+        }
+    }
+
+    $CacheKey .= '-' . $Self->{SortBy} if defined $Self->{SortBy};
+
+    # Set OrderBy parameter to the search.
+    my $IsCacheForUse = 0;
+    if ( $Self->{OrderBy} ) {
+        if (
+            $Self->{AddFilters}
+            || $Self->{TabAction}
+            || !defined $Self->{SortingColumn}
+            || $Self->{SortingColumn} ne $Self->{SortBy}
+            )
+        {
+            $TicketSearch{OrderBy} = $Self->{OrderBy};
+            $IsCacheForUse = 1;
+        }
+        else {
+            $TicketSearch{OrderBy} = $Self->{OrderBy} eq 'Up' ? 'Down' : 'Up';
+        }
+    }
+
+    # Set previous sorting column parameter for all columns.
+    $Param{SortingColumn} = $Self->{SortBy};
+
+    $CacheKey .= '-' . $TicketSearch{OrderBy} if defined $TicketSearch{OrderBy};
 
     # CustomerInformationCenter shows data per CustomerID
     if ( $Param{CustomerID} ) {
@@ -629,10 +668,7 @@ sub Run {
             }
         }
 
-        # add order by parameter to the search
-        if ( $Self->{OrderBy} ) {
-            $TicketSearch{OrderBy} = $Self->{OrderBy};
-        }
+        $CacheUsed = $IsCacheForUse ? 1 : 0;
 
         # add process management search terms
         if ( $Self->{Config}->{IsProcessWidget} ) {
@@ -641,7 +677,6 @@ sub Run {
             };
         }
 
-        $CacheUsed = 0;
         my @TicketIDsArray;
         if ( !$Self->{Config}->{IsProcessWidget} || IsArrayRefWithData( $Self->{ProcessList} ) ) {
 
@@ -945,7 +980,7 @@ sub Run {
         . ';Filter=' . $Self->{Filter}
         . ';AdditionalFilter=' . ( $Self->{AdditionalFilter} || '' )
         . ';SortBy=' .           ( $Self->{SortBy}           || '' )
-        . ';OrderBy=' .          ( $Self->{OrderBy}          || '' )
+        . ';OrderBy=' .          ( $TicketSearch{OrderBy}    || '' )
         . $ColumnFilterLink
         . ';';
 
@@ -955,6 +990,7 @@ sub Run {
     if ( $Param{CustomerUserID} ) {
         $LinkPage .= "CustomerUserID=$Param{CustomerUserID};";
     }
+
     my %PageNav = $LayoutObject->PageNavBar(
         StartHit    => $Self->{StartHit},
         PageShown   => $Self->{PageShown},
@@ -989,22 +1025,35 @@ sub Run {
 
     # show non-labeled table headers
     my $CSS = '';
-    my $OrderBy;
+
+    # Set order for blocks.
+    $TicketSearch{OrderBy} = $TicketSearch{OrderBy} || 'Up';
+
+    # Send data to JS for init container.
+    $LayoutObject->AddJSData(
+        Key   => 'InitContainerDashboard' . $Self->{Name},
+        Value => {
+            SortBy => $Self->{SortBy} || 'Up',
+            OrderBy => $TicketSearch{OrderBy},
+        },
+    );
+
     for my $Item (@MetaItems) {
         $CSS = '';
-        my $Title = $Item;
-        if ( $Self->{SortBy} && ( $Self->{SortBy} eq $Item ) ) {
-            if ( $Self->{OrderBy} && ( $Self->{OrderBy} eq 'Up' ) ) {
-                $OrderBy = 'Down';
+        my $Title = $LayoutObject->{LanguageObject}->Translate($Item);
+
+        # set title description
+        if ( $Self->{SortBy} && $Self->{SortBy} eq $Item ) {
+            my $TitleDesc = '';
+            if ( $TicketSearch{OrderBy} eq 'Down' ) {
                 $CSS .= ' SortAscendingLarge';
+                $TitleDesc = Translatable('sorted descending');
             }
             else {
-                $OrderBy = 'Up';
                 $CSS .= ' SortDescendingLarge';
+                $TitleDesc = Translatable('sorted ascending');
             }
 
-            # set title description
-            my $TitleDesc = $OrderBy eq 'Down' ? Translatable('sorted ascending') : Translatable('sorted descending');
             $TitleDesc = $LayoutObject->{LanguageObject}->Translate($TitleDesc);
             $Title .= ', ' . $TitleDesc;
         }
@@ -1038,7 +1087,7 @@ sub Run {
                 Key   => 'HeaderMeta' . $WidgetName,
                 Value => {
                     Name             => $Self->{Name},
-                    OrderBy          => $OrderBy || 'Up',
+                    OrderBy          => $TicketSearch{OrderBy},
                     HeaderColumnName => $Item,
                 },
             );
@@ -1048,7 +1097,7 @@ sub Run {
                 Data => {
                     %Param,
                     Name             => $Self->{Name},
-                    OrderBy          => $OrderBy || 'Up',
+                    OrderBy          => $TicketSearch{OrderBy},
                     HeaderColumnName => $Item,
                     Title            => $Title,
                 },
@@ -1076,19 +1125,18 @@ sub Run {
             $CSS = '';
             my $Title = $LayoutObject->{LanguageObject}->Translate($HeaderColumn);
 
-            if ( $Self->{SortBy} && ( $Self->{SortBy} eq $HeaderColumn ) ) {
-                if ( $Self->{OrderBy} && ( $Self->{OrderBy} eq 'Up' ) ) {
-                    $OrderBy = 'Down';
+            # Set title description.
+            if ( $Self->{SortBy} && $Self->{SortBy} eq $HeaderColumn ) {
+                my $TitleDesc = '';
+                if ( $TicketSearch{OrderBy} eq 'Down' ) {
                     $CSS .= ' SortAscendingLarge';
+                    $TitleDesc = Translatable('sorted descending');
                 }
                 else {
-                    $OrderBy = 'Up';
                     $CSS .= ' SortDescendingLarge';
+                    $TitleDesc = Translatable('sorted ascending');
                 }
 
-                # add title description
-                my $TitleDesc
-                    = $OrderBy eq 'Down' ? Translatable('sorted ascending') : Translatable('sorted descending');
                 $TitleDesc = $LayoutObject->{LanguageObject}->Translate($TitleDesc);
                 $Title .= ', ' . $TitleDesc;
             }
@@ -1135,8 +1183,9 @@ sub Run {
                 $LayoutObject->AddJSData(
                     Key   => 'TicketNumberColumn' . $WidgetName,
                     Value => {
-                        Name    => $Self->{Name},
-                        OrderBy => $OrderBy || 'Up',
+                        Name          => $Self->{Name},
+                        OrderBy       => $TicketSearch{OrderBy},
+                        SortingColumn => $Param{SortingColumn},
                     },
                 );
 
@@ -1200,7 +1249,8 @@ sub Run {
                         HeaderColumnName => $HeaderColumn,
                         Name             => $Self->{Name},
                         SortBy           => $Self->{SortBy} || 'Age',
-                        OrderBy          => $OrderBy || 'Up',
+                        OrderBy          => $TicketSearch{OrderBy},
+                        SortingColumn    => $Param{SortingColumn},
                     },
                 );
 
@@ -1212,7 +1262,7 @@ sub Run {
                         CSS                  => $CSS,
                         HeaderNameTranslated => $TranslatedWord || $HeaderColumn,
                         ColumnFilterStrg     => $ColumnFilterHTML,
-                        OrderBy              => $OrderBy || 'Up',
+                        OrderBy              => $TicketSearch{OrderBy},
                         SortBy               => $Self->{SortBy} || 'Age',
                         Name                 => $Self->{Name},
                         Title                => $Title,
@@ -1276,7 +1326,7 @@ sub Run {
                         HeaderColumnName => $HeaderColumn,
                         Name             => $Self->{Name},
                         SortBy           => $Self->{SortBy} || 'Age',
-                        OrderBy          => $OrderBy || 'Up',
+                        OrderBy          => $TicketSearch{OrderBy},
                     },
                 );
 
@@ -1322,7 +1372,8 @@ sub Run {
                         HeaderColumnName => $HeaderColumn,
                         Name             => $Self->{Name},
                         SortBy           => $Self->{SortBy} || $HeaderColumn,
-                        OrderBy          => $OrderBy || 'Up',
+                        OrderBy          => $TicketSearch{OrderBy},
+                        SortingColumn    => $Param{SortingColumn},
                     },
                 );
 
@@ -1333,7 +1384,7 @@ sub Run {
                         HeaderColumnName     => $HeaderColumn,
                         CSS                  => $CSS,
                         HeaderNameTranslated => $TranslatedWord || $HeaderColumn,
-                        OrderBy              => $OrderBy || 'Up',
+                        OrderBy              => $TicketSearch{OrderBy},
                         SortBy               => $Self->{SortBy} || $HeaderColumn,
                         Name                 => $Self->{Name},
                         Title                => $Title,
@@ -1404,24 +1455,21 @@ sub Run {
             );
 
             if ($IsSortable) {
-                my $OrderBy;
+                my $TitleDesc = '';
                 if (
                     $Self->{SortBy}
                     && ( $Self->{SortBy} eq ( 'DynamicField_' . $DynamicFieldConfig->{Name} ) )
                     )
                 {
-                    if ( $Self->{OrderBy} && ( $Self->{OrderBy} eq 'Up' ) ) {
-                        $OrderBy = 'Down';
+                    if ( $TicketSearch{OrderBy} eq 'Down' ) {
                         $CSS .= ' SortAscendingLarge';
+                        $TitleDesc = Translatable('sorted descending');
                     }
                     else {
-                        $OrderBy = 'Up';
                         $CSS .= ' SortDescendingLarge';
+                        $TitleDesc = Translatable('sorted ascending');
                     }
 
-                    # add title description
-                    my $TitleDesc
-                        = $OrderBy eq 'Down' ? Translatable('sorted ascending') : Translatable('sorted descending');
                     $TitleDesc = $LayoutObject->{LanguageObject}->Translate($TitleDesc);
                     $Title .= ', ' . $TitleDesc;
                 }
@@ -1450,7 +1498,8 @@ sub Run {
                             HeaderColumnName => $DynamicFieldName,
                             Name             => $Self->{Name},
                             SortBy           => $Self->{SortBy} || 'Age',
-                            OrderBy          => $OrderBy || 'Up',
+                            OrderBy          => $TicketSearch{OrderBy},
+                            SortingColumn    => $Param{SortingColumn},
                         },
                     );
 
@@ -1463,7 +1512,7 @@ sub Run {
                             CSS                  => $CSS,
                             HeaderNameTranslated => $TranslatedLabel || $DynamicFieldName,
                             ColumnFilterStrg     => $ColumnFilterHTML,
-                            OrderBy              => $OrderBy || 'Up',
+                            OrderBy              => $TicketSearch{OrderBy},
                             SortBy               => $Self->{SortBy} || 'Age',
                             Name                 => $Self->{Name},
                             Title                => $Title,
@@ -1482,7 +1531,8 @@ sub Run {
                             HeaderColumnName => $DynamicFieldName,
                             Name             => $Self->{Name},
                             SortBy           => $Self->{SortBy} || $DynamicFieldName,
-                            OrderBy          => $OrderBy || 'Up',
+                            OrderBy          => $TicketSearch{OrderBy},
+                            SortingColumn    => $Param{SortingColumn},
                         },
                     );
 
@@ -1494,7 +1544,7 @@ sub Run {
                             HeaderColumnName     => $DynamicFieldName,
                             CSS                  => $CSS,
                             HeaderNameTranslated => $TranslatedLabel || $DynamicFieldName,
-                            OrderBy              => $OrderBy || 'Up',
+                            OrderBy              => $TicketSearch{OrderBy},
                             SortBy               => $Self->{SortBy} || $DynamicFieldName,
                             Name                 => $Self->{Name},
                             Title                => $Title,
@@ -1530,7 +1580,7 @@ sub Run {
                         HeaderColumnName => $DynamicFieldName,
                         Name             => $Self->{Name},
                         SortBy           => $Self->{SortBy} || 'Age',
-                        OrderBy          => $OrderBy || 'Up',
+                        OrderBy          => $TicketSearch{OrderBy},
                     },
                 );
 
@@ -1946,6 +1996,9 @@ sub Run {
             CustomerID     => $Param{CustomerID},
             CustomerUserID => $Param{CustomerUserID},
             FilterActive   => $FilterActive,
+            SortBy         => $Self->{SortBy} || 'Age',
+            OrderBy        => $TicketSearch{OrderBy},
+            SortingColumn  => $Param{SortingColumn},
         },
     );